requirements.txt
wheel
nonebot-plugin-apscheduler
nonebot-adapter-onebot
requests
pytz==2023.3.post1
humanize==4.8.0
pillow==9.5.0
nonebot-plugin-imageutils
nonebot-plugin-admin
nonebot-plugin-cloudsignx
websocket-client